dts: driver: bindings: Add support for Clock-control for display
Add clock-control support for display to remove clock calculation and
setting in hardware inconsistencies.

Also added bindings in dts for max lane bandwidth support.

+++ include/zephyr/drivers/mipi_dphy/dphy_dw.h
@@ -17,7 +17,7 @@ extern "C" {
 
 /* Utility functions. */
 #define CEIL(x) ((x) > (uint32_t)(x) ? (uint32_t) (x+1) : (uint32_t) (x))
-#define ROUND(x) (uint32_t)((x) + 0.5f)
+#define ROUND(x) (uint32_t)((double)(x) + 0.5)
 
 struct dphy_dsi_settings {
 	/* Number of lanes in D-PHY. */
